+++ pioreactor/automations/dosing/chemostat.py
@@ -32,6 +32,6 @@ class Chemostat(DosingAutomationJob):
     def execute(self) -> events.DilutionEvent:
         volume_actually_cycled = self.execute_io_action(media_ml=self.volume, waste_ml=self.volume)
         return events.DilutionEvent(
-            f"exchanged {volume_actually_cycled[0]}mL",
-            data={"volume_actually_cycled": volume_actually_cycled[0]},
+            f"exchanged {volume_actually_cycled['waste_ml']}mL",
+            data={"volume_actually_cycled": volume_actually_cycled["waste_ml"]},
         )